+++ e2e/__tests__/hasteMapSize.test.ts
@@ -7,11 +7,11 @@
 
 import {tmpdir} from 'os';
 import * as path from 'path';
+import {realpathSync} from 'graceful-fs';
 import HasteMap = require('jest-haste-map');
-import {sync as realpath} from 'realpath-native';
 import {cleanup, writeFiles} from '../Utils';
 
-const DIR = path.resolve(realpath(tmpdir()), 'haste_map_size');
+const DIR = path.resolve(realpathSync.native(tmpdir()), 'haste_map_size');
 
 beforeEach(() => {
   cleanup(DIR);