+++ packages/jest-resolve/src/defaultResolver.ts
@@ -7,8 +7,8 @@
 
 import * as fs from 'graceful-fs';
 import {sync as resolveSync} from 'resolve';
-import {sync as realpath} from 'realpath-native';
 import pnpResolver from 'jest-pnp-resolver';
+import {tryRealpath} from 'jest-util';
 import type {Config} from '@jest/types';
 
 type ResolverOptions = {
@@ -95,17 +95,7 @@ function realpathCached(path: Config.Path): Config.Path {
     return result;
   }
 
-  try {
-    result = realpath(path);
-  } catch (error) {
-    if (error.code !== 'ENOENT') {
-      throw error;
-    }
-  }
-
-  if (!result) {
-    result = path;
-  }
+  result = tryRealpath(path);
 
   checkedRealpathPaths.set(path, result);