+++ packages/jest-resolve/src/nodeModulesPaths.ts
@@ -9,7 +9,7 @@
 
 import * as path from 'path';
 import type {Config} from '@jest/types';
-import {sync as realpath} from 'realpath-native';
+import {tryRealpath} from 'jest-util';
 
 type NodeModulesPathsOptions = {
   moduleDirectory?: Array<string>;
@@ -40,7 +40,7 @@ export default function nodeModulesPaths(
   // traverses parents of the physical path, not the symlinked path
   let physicalBasedir;
   try {
-    physicalBasedir = realpath(basedirAbs);
+    physicalBasedir = tryRealpath(basedirAbs);
   } catch (err) {
     // realpath can throw, e.g. on mapped drives
     physicalBasedir = basedirAbs;