+++ packages/jest-runtime/src/cli/index.ts
@@ -8,12 +8,11 @@
 import {cpus} from 'os';
 import * as path from 'path';
 import chalk = require('chalk');
-import {sync as realpath} from 'realpath-native';
 import yargs = require('yargs');
 import type {Config} from '@jest/types';
 import type {JestEnvironment} from '@jest/environment';
 import {CustomConsole} from '@jest/console';
-import {setGlobal} from 'jest-util';
+import {setGlobal, tryRealpath} from 'jest-util';
 import {validateCLIOptions} from 'jest-validate';
 import {deprecationEntries, readConfig} from 'jest-config';
 import {VERSION} from '../version';
@@ -53,7 +52,7 @@ export async function run(
     return;
   }
 
-  const root = realpath(process.cwd());
+  const root = tryRealpath(process.cwd());
   const filePath = path.resolve(root, argv._[0]);
 
   if (argv.debug) {